+++ src/Bozboz/Admin/Fields/CheckboxField.php
@@ -6,6 +6,18 @@ class CheckboxField extends Field
 {
 	public function getInput()
 	{
-		return Form::hidden($this->get('name'), 0, ['id' => 'hidden_' . $this->get('name')]) . Form::checkbox($this->get('name'), 1);
+		return '<input type="hidden" name="' . $this->get('name') . '" value="">'
+		     . Form::checkbox($this->get('name'), 1, $this->getCheckedState());
+	}
+
+	/**
+	 * Get the checked state, or fall back to session/model data if nothing is
+	 * set (null returned)
+	 *
+	 * @return boolean|null
+	 */
+	protected function getCheckedState()
+	{
+		return $this->get('checked');
 	}
 }