+++ src/Bozboz/Admin/Models/User.php
@@ -137,4 +137,14 @@ class User extends Base implements UserInterface, RemindableInterface, Permissio
 			});
 		});
 	}
+
+	public function scopeDoesntHavePermission($builder, $action)
+	{
+		$builder->whereDoesntHave('permissions', function($q) use ($action) {
+			$q->where(function($q) use ($action) {
+				$q->where('action', $action)
+				  ->orWhere('action', Permission::WILDCARD);
+			});
+		});
+	}
 }