+++ src/Bozboz/Admin/Permissions/PermissionAdminDecorator.php
@@ -42,6 +42,8 @@ class PermissionAdminDecorator extends ModelAdminDecorator
 	{
 		$rules = array_keys($this->permissions->dump());
 
+		array_unshift($rules, Permission::WILDCARD);
+
 		return array_combine($rules, $rules);
 	}
 
@@ -77,7 +79,18 @@ class PermissionAdminDecorator extends ModelAdminDecorator
 	public function getListingFilters()
 	{
 		return [
-			new ArrayListingFilter('user', ['' => 'All'] + $this->model->user()->getModel()->lists('first_name', 'id'), 'user_id')
+			new ArrayListingFilter('user', $this->getListOfAdminUsers(), 'user_id')
 		];
+	}
+
+	protected function getListOfAdminUsers()
+	{
+		$users = $this->users->getListOfAdminUsers()->keyBy('id')->map(function($user) {
+			return $this->users->getLabel($user);
+		});
+
+		$users->prepend('All', '');
+
+		return $users;
 	}
 }